IGNITE-21750 / move a method
